fix: send environment on props of component
